third_party/blink/web_tests/external/wpt/editing/other/formatblock-preserving-selection.tentative_styleWithCSS=false_block=address-expected.txt

This is a testharness.js-based test.
[FAIL] Preserve selection after formatBlock with address at <table><tr><td>a[bc</td></tr></table><div>de]f</div>
  assert_equals: Result: <address><table><tbody><tr><td><address>abc</address></td></tr></tbody></table>def</address> expected "bcde" but got "bcdef"
Harness: the test ran to completion.
